1.系统消息显示不全的 bug 2.时间轴并发崩溃 bug

+++ app/src/main/java/com/cylan/jiafeigou/dp/DpMsgDefine.java
@@ -2174,7 +2174,7 @@ public class DpMsgDefine {
         @Index(2)
         public String faceId;
         @Index(3)
-        public int timeMsec;
+        public long timeMsec;
         @Index(4)
         public List<DPHeader> dataList;
